extract file naming expectation from requantify and summarize_data + remove execution over multiple files of summarize_data
