+++ easy_fuse/processing.py
@@ -15,6 +15,7 @@ import time
 from shutil import copy
 
 import logzero
+from easy_fuse.misc.count_input_reads import get_input_read_count
 from logzero import logger
 
 import easy_fuse
@@ -33,20 +34,25 @@ class Processing(object):
 
     def __init__(
         self,
-        cmd,
-        input_paths,
+        sample_id: str,
+        fastq1: str,
+        fastq2: str,
         working_dir,
         config: EasyFuseConfiguration,
         jobname_suffix,
     ):
         """Parameter initiation and work folder creation. Start of progress logging."""
-        self.working_dir = os.path.abspath(working_dir)
+        self.working_dir = os.path.join(os.path.abspath(working_dir), sample_id)
         self.log_path = os.path.join(self.working_dir, "easyfuse_processing.log")
         logzero.logfile(self.log_path)
         io_methods.create_folder(self.working_dir)
 
-        logger.info("Starting easyfuse: CMD - {}".format(cmd))
-        self.input_paths = [os.path.abspath(file) for file in input_paths]
+        logger.info("Starting easyfuse!")
+        self.sample_id = sample_id
+        assert os.path.exists(fastq1), "Fastq file {} does not exist!".format(fastq1)
+        assert os.path.exists(fastq2), "Fastq file {} does not exist!".format(fastq2)
+        self.fastq1 = fastq1
+        self.fastq2 = fastq2
 
         self.cfg = config
         copy(self.cfg.config_file, working_dir)
@@ -70,44 +76,23 @@ class Processing(object):
             )
         )
 
-        # get fastq files
-        fastqs = io_methods.get_fastq_files(self.input_paths)
-        sample_list = io_methods.pair_fastq_files(fastqs)
-        for sample_id, fq1, fq2 in sample_list:
-            if fq1 and fq2:
-                logger.info("Processing Sample ID: {} (paired end)".format(sample_id))
-                logger.info("Sample 1: {}".format(fq1))
-                logger.info("Sample 2: {}".format(fq2))
-                self.execute_pipeline(fq1, fq2, sample_id, ref_genome, ref_trans)
+        self.execute_pipeline(self.fastq1, self.fastq2, self.sample_id, ref_genome, ref_trans)
 
         # summarize all data if selected
         if "summary" in self.cfg["general"]["tools"].split(","):
             dependency = []
-            for sample in sample_list:
-                dependency.extend(
-                    queueing.get_jobs_by_name(
-                        "requantifyBest-{}".format(sample[0]),
-                        self.cfg["general"]["queueing_system"],
-                    )
+
+            dependency.extend(
+                queueing.get_jobs_by_name(
+                    "requantifyBest-{}".format(self.sample_id),
+                    self.cfg["general"]["queueing_system"],
                 )
+            )
 
             # TODO: add support for running without model and multiple models
-            modelling_string = " --model_predictions"
 
-            cmd_summarize = (
-                "easy-fuse summarize-data --input {0}{1} -c {2} --samples {3}".format(
-                    self.working_dir,
-                    modelling_string,
-                    self.cfg.config_file,
-                    " ".join([sample_id for sample_id, _, _ in sample_list]),
-                )
-            )
+            cmd_summarize = self._build_command_summarize_data()
 
-            logger.debug(
-                "Submitting job: CMD - {0}; PATH - {1}; DEPS - {2}".format(
-                    cmd_summarize, self.working_dir, dependency
-                )
-            )
             resources_summary = self.cfg["resources"]["summary"].split(",")
             cpu = resources_summary[0]
             mem = resources_summary[1]
@@ -126,6 +111,65 @@ class Processing(object):
                 dependency,
                 self.cfg["general"]["receiver"],
             )
+
+    def _build_command_summarize_data(self):
+        # summarize data input files
+        context_seq_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "fetched_contextseqs",
+            "Context_Seqs.csv",
+        )
+        detect_fusion_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "fetched_fusions",
+            "Detected_Fusions.csv",
+        )
+        requant_mode = self.cfg["general"]["requant_mode"]
+        requant_cpm_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "classification",
+            "classification_{}.tdt".format(requant_mode),
+        )
+        requant_cnt_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "classification",
+            "classification_{}.tdt.counts".format(requant_mode),
+        )
+        input_reads_stats_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "classification",
+            "Star_org_input_reads.txt",
+        )
+        # summarize data output folder
+        fusion_summary_output_folder = os.path.join(self.working_dir, "fusion_summary")
+        io_methods.create_folder(fusion_summary_output_folder)
+        cmd_summarize = (
+            "easy-fuse summarize-data "
+            "--input-fusions {input_fusions} "
+            "--input-fusion-context-seqs {input_fusion_context_seqs} "
+            "--input-requant-cpm {input_requant_cpm} "
+            "--input-requant-counts {input_requant_counts} "
+            "--input-reads-stats {input_reads_stats} "
+            "--model_predictions "
+            "--output-folder {output_folder} "
+            "--config-file {config} "
+            "--count-reads {count_reads}".format(
+                input_fusions=detect_fusion_file,
+                input_fusion_context_seqs=context_seq_file,
+                input_requant_cpm=requant_cpm_file,
+                input_requant_counts=requant_cnt_file,
+                input_reads_stats=input_reads_stats_file,
+                output_folder=fusion_summary_output_folder,
+                config=self.cfg.config_file,
+                count_reads=self.count_reads
+            )
+        )
+        return cmd_summarize
 
     # Per sample, define input parameters and execution commands, create a folder tree and submit runs to slurm
     def execute_pipeline(self, fq1, fq2, sample_id, ref_genome, ref_trans):
@@ -153,22 +197,19 @@ class Processing(object):
         # Output results folder creation - currently included:
         # 1) Gene/Isoform expression: star
         # 2) Fusion prediction: mapsplice, fusioncatcher, starfusion, infusion, soapfuse
-        output_results_path = os.path.join(
-            self.working_dir, "Sample_{}".format(sample_id)
-        )
-        qc_path = os.path.join(output_results_path, "qc")
+        qc_path = os.path.join(self.working_dir, "qc")
         skewer_path = os.path.join(qc_path, "skewer")
         qc_table_path = os.path.join(qc_path, "qc_table.txt")
-        filtered_reads_path = os.path.join(output_results_path, "filtered_reads")
-        expression_path = os.path.join(output_results_path, "expression")
+        filtered_reads_path = os.path.join(self.working_dir, "filtered_reads")
+        expression_path = os.path.join(self.working_dir, "expression")
         star_path = os.path.join(expression_path, "star")
-        fusion_path = os.path.join(output_results_path, "fusion")
+        fusion_path = os.path.join(self.working_dir, "fusion")
         mapsplice_path = os.path.join(fusion_path, "mapsplice")
         fusioncatcher_path = os.path.join(fusion_path, "fusioncatcher")
         starfusion_path = os.path.join(fusion_path, "starfusion")
         infusion_path = os.path.join(fusion_path, "infusion")
         soapfuse_path = os.path.join(fusion_path, "soapfuse")
-        fetchdata_path = os.path.join(output_results_path, "fetchdata")
+        fetchdata_path = os.path.join(self.working_dir, "fetchdata")
         # former fetchdata.py
         detected_fusions_path = os.path.join(fetchdata_path, "fetched_fusions")
         detected_fusions_file = os.path.join(
@@ -195,7 +236,6 @@ class Processing(object):
         )
 
         for folder in [
-            output_results_path,
             qc_path,
             skewer_path,
             filtered_reads_path,
@@ -380,19 +420,6 @@ class Processing(object):
             )
         )
 
-        # (8) Data collection
-        # TODO: embed this as operations within the package
-        cmd_readcounts = (
-            "easy-fuse count-reads "
-            "-i {0} "
-            "-f {1} "
-            "-o {2}".format(
-                os.path.join(filtered_reads_path, "{}_Log.final.out".format(sample_id)),
-                "star",
-                os.path.join(classification_path, "Star_org_input_reads.txt"),
-            )
-        )
-
         cmd_fusiondata = (
             "easy-fuse fusion-parser "
             "-i {0} "
@@ -400,7 +427,7 @@ class Processing(object):
             "-s {2} "
             "-f {3} "
             "-l {4}".format(
-                output_results_path,
+                self.working_dir,
                 detected_fusions_path,
                 sample_id,
                 self.cfg["general"]["fusiontools"],
@@ -447,11 +474,18 @@ class Processing(object):
             cmds["samtools"], star_align_file
         )
 
+        input_reads_stats_file = os.path.join(
+            self.working_dir,
+            "fetchdata",
+            "classification",
+            "Star_org_input_reads.txt",
+        )
         cmd_requantify_fltr = (
             "easy-fuse requantify "
             "-i {0}fltr_Aligned.sortedByCoord.out.bam "
             "-o {1}_fltr.tdt "
-            "-d 10".format(star_align_file, classification_file)
+            "-d 10 "
+            "--input-reads-stats {2}".format(star_align_file, classification_file, input_reads_stats_file)
         )
 
         cmd_staralign_org = (
@@ -471,7 +505,9 @@ class Processing(object):
         cmd_requantify_org = (
             "easy-fuse requantify "
             "-i {0}org_Aligned.sortedByCoord.out.bam "
-            "-o {1}_org.tdt -d 10".format(star_align_file, classification_file)
+            "-o {1}_org.tdt "
+            "-d 10 "
+            "--input-reads-stats {2}".format(star_align_file, classification_file, input_reads_stats_file)
         )
 
         # for testing, based on debug. should be removed if merged to original
@@ -534,7 +570,9 @@ class Processing(object):
         cmd_requantify_best = (
             "easy-fuse requantify "
             "-i {0}best_Aligned.sortedByCoord.out.bam "
-            "-o {1}_best.tdt -d 10".format(star_align_file, classification_file)
+            "-o {1}_best.tdt "
+            "-d 10"
+            "--input-reads-stats {2}".format(star_align_file, classification_file, input_reads_stats_file)
         )
 
         # set final lists of executable tools and path
@@ -572,7 +610,6 @@ class Processing(object):
             cmd_starfusion,
             cmd_infusion,
             cmd_soapfuse,
-            cmd_readcounts,
             cmd_fusiondata,
             cmd_contextseq,
             cmd_starindex,
@@ -679,12 +716,6 @@ class Processing(object):
 
                 # Build slurm dependencies
                 dependencies = self.build_dependencies(tool, sample_id)
-
-                logger.debug(
-                    "Submitting job: CMD - {0}; PATH - {1}; DEPS - {2}".format(
-                        cmd, exe_path[i], dependencies
-                    )
-                )
                 self.submit_job(uid, cmd, cpu, mem, exe_path[i], dependencies, "")
             else:
                 logger.info(
@@ -697,6 +728,11 @@ class Processing(object):
         self, uid, cmd, cores, mem_usage, output_results_folder, dependencies, mail
     ):
         """Submit job to for process generation"""
+        logger.debug(
+            "Submitting job: CMD - {0}; PATH - {1}; DEPS - {2}".format(
+                cmd, output_results_folder, dependencies
+            )
+        )
         que_sys = self.cfg["general"]["queueing_system"]
         already_running = queueing.get_jobs_by_name(uid, que_sys)
         if not already_running:
@@ -787,11 +823,24 @@ class Processing(object):
 
 def add_pipeline_parser_args(pipeline_parser):
     pipeline_parser.add_argument(
-        "-i",
-        "--input",
-        dest="input_paths",
-        nargs="+",
-        help="Specify full path of the fastq folder to process.",
+        "-1",
+        "--fastq1",
+        dest="fastq1",
+        help="FASTQ file for read 1",
+        required=True,
+    )
+    pipeline_parser.add_argument(
+        "-2",
+        "--fastq2",
+        dest="fastq2",
+        help="FASTQ file for read 2",
+        required=True,
+    )
+    pipeline_parser.add_argument(
+        "-s",
+        "--sample-id",
+        dest="sample_id",
+        help="Sample identifier",
         required=True,
     )
     pipeline_parser.add_argument(
@@ -847,5 +896,5 @@ def pipeline_command(args):
         outf.write("#!/bin/sh\n\n{}".format(script_call))
 
     Processing(
-        script_call, args.input_paths, args.output_folder, config, args.jobname_suffix
+        args.sample_id, args.fastq1, args.fastq2, args.output_folder, config, args.jobname_suffix
     ).run()
